# perf: use dirtmake to reduce memclr cost