--- a/pkg/generic/thrift/json.go
+++ b/pkg/generic/thrift/json.go
@@ -22,6 +22,7 @@ import (
 	"strconv"
 
 	"github.com/apache/thrift/lib/go/thrift"
+	"github.com/bytedance/gopkg/lang/dirtmake"
 	"github.com/cloudwego/dynamicgo/conv"
 	"github.com/cloudwego/dynamicgo/conv/t2j"
 	dthrift "github.com/cloudwego/dynamicgo/thrift"
@@ -204,7 +205,7 @@ func (m *ReadJSON) Read(ctx context.Context, method string, in thrift.TProtocol)
 		}
 
 		// json size is usually 2 times larger than equivalent thrift data
-		buf := make([]byte, 0, len(transBuff)*2)
+		buf := dirtmake.Bytes(0, len(transBuff)*2)
 		// thrift []byte to json []byte
 		if err := m.t2jBinaryConv.DoInto(ctx, tyDsc, transBuff, &buf); err != nil {
 			return nil, err