--- a/pkg/remote/trans/nphttp2/server_conn.go
+++ b/pkg/remote/trans/nphttp2/server_conn.go
@@ -22,6 +22,8 @@ import (
 	"net"
 	"time"
 
+	"github.com/bytedance/gopkg/lang/dirtmake"
+
 	"github.com/cloudwego/kitex/pkg/remote/trans/nphttp2/codes"
 	"github.com/cloudwego/kitex/pkg/remote/trans/nphttp2/grpc"
 	"github.com/cloudwego/kitex/pkg/remote/trans/nphttp2/status"
@@ -43,13 +45,13 @@ func newServerConn(tr grpc.ServerTransport, s *grpc.Stream) *serverConn {
 }
 
 func (c *serverConn) ReadFrame() (hdr, data []byte, err error) {
-	hdr = make([]byte, 5)
+	hdr = dirtmake.Bytes(5, 5)
 	_, err = c.Read(hdr)
 	if err != nil {
 		return nil, nil, err
 	}
 	dLen := int(binary.BigEndian.Uint32(hdr[1:]))
-	data = make([]byte, dLen)
+	data = dirtmake.Bytes(dLen, dLen)
 	_, err = c.Read(data)
 	if err != nil {
 		return nil, nil, err