@jayantxie 2024-05-22
**Before：**
```
~/> ./scripts/benchmark_thrift.sh
Checking whether the environment meets the requirements ...
Check finished.
Building thrift services by exec build_thrift.sh ...
Build finished.
Starting server [kitex], if failed please check [output/log/nohup.log] for detail
Server [kitex] running with [ numactl -C 0,1,2,3 -m 0]
Client [kitex] running with [ numactl -C 12,13,14,15,16,17,18,19,20,21,22,23,36,37,38,39 -m 1]
Info: [KITEX] start benching [2024-05-22 20:08:48.221844203 +0800 CST m=+0.797783407], concurrent: 100, total: 3000000, sleep: 0
Info: [KITEX]: finish benching [2024-05-22 20:09:11.351526226 +0800 CST m=+23.927465435], took 23129 ms for 3000000 requests
Info: [KITEX]: requests total: 3000000, failed: 0
Info: [KITEX]: TPS: 129703.82, TP99: 2.17ms, TP999: 11.16ms (b=10240 Byte, c=100, n=3000000)
[KITEX@Server] CPU Usage: MIN: 373.83%, TP50: 388.67%, TP90: 391.96%, TP99: 392.39%, MAX: 392.61%, AVG:388.84%
[KITEX@Server] Mem Usage: AVG: 67 MB, MAX: 71 MB
[KITEX@Client] CPU Usage: MIN: 498.00%, TP50: 507.51%, TP90: 511.00%, TP99: 514.85%, MAX: 515.52%, AVG:507.32%
[KITEX@Client] Mem Usage: AVG: 93 MB, MAX: 133 MB
```

**After：**
```
tiger@n129-036-075:~/xiezhengyao/github.com/cloudwego/kitex-benchmark$ ./scripts/benchmark_thrift.sh
Checking whether the environment meets the requirements ...
Check finished.
Building thrift services by exec build_thrift.sh ...
Build finished.
Starting server [kitex], if failed please check [output/log/nohup.log] for detail
Server [kitex] running with [ numactl -C 0,1,2,3 -m 0]
Client [kitex] running with [ numactl -C 12,13,14,15,16,17,18,19,20,21,22,23,36,37,38,39 -m 1]
Info: [KITEX] start benching [2024-05-22 20:09:50.87681807 +0800 CST m=+0.763075252], concurrent: 100, total: 3000000, sleep: 0
Info: [KITEX]: finish benching [2024-05-22 20:10:12.58053603 +0800 CST m=+22.466793212], took 21703 ms for 3000000 requests
Info: [KITEX]: requests total: 3000000, failed: 0
Info: [KITEX]: TPS: 138225.49, TP99: 2.10ms, TP999: 11.60ms (b=10240 Byte, c=100, n=3000000)
[KITEX@Server] CPU Usage: MIN: 366.34%, TP50: 390.17%, TP90: 393.18%, TP99: 393.33%, MAX: 393.39%, AVG:389.48%
[KITEX@Server] Mem Usage: AVG: 72 MB, MAX: 78 MB
[KITEX@Client] CPU Usage: MIN: 445.00%, TP50: 510.00%, TP90: 515.00%, TP99: 516.00%, MAX: 518.89%, AVG:507.72%
[KITEX@Client] Mem Usage: AVG: 86 MB, MAX: 93 MB
```